# Speed up railties test test_skip_webpack_install currently ~10s