### Summary

There is another test in the generators that is taking ~10s: `AppGeneratorTest#test_skip_webpack_install = 9.79s`

To repo: `b exec ruby -w -Itest test/generators/app_generator_test.rb -v  -n test_skip_webpack_install`

I have removed the `rails --help` as it does not seem to be part of generators, which reduced this test to ~5s. The `skip_bundle` reduced this test by a further 4s.

Related to: https://github.com/rails/rails/pull/39661