@eugeneius 2020-06-29
Thanks! Saving 5 seconds here is pretty cool 🐎